Fixed lag in the list with big images.
Instead of loading the images in the delegate they are now loaded and
resized before setting the model items. The thumbnails are then passed
to the delegate.

A few things to keep in mind:
- The app now uses more memory because all the images are loaded (this
can be be improved by not loading all the images at once but on demand).
- The initial loading is now longer, same for changing layout types. The
app is irresponsive while loading, it would be better if we change the
loading to a bakcground thread.